Back to softmax, new temperatured prediction model.
We now use a separate prediction model which is almost identical, except that
we do not have any dropout layers and instead have a lambda layer which
modifies the input to the final softmax layer to implement a temperature
setting. The generated text is sampled from a distribution defined by the
output of the softmax layer, and the temperature parameter allows us to make
this distribution more even or more pronounced towards safer outputs.